C - BinTrees - FIX: "increment" type for exos 9 to 13.
Just realized (finally) when initializing 14th that
  my functions returned "unsigned int" whereas
  prototype said "size_t".
Since it worked I guess there was some on the fly
  conversion but it's better to just be aligned
  from the get go.

+++ 10-binary_tree_depth.c
@@ -13,7 +13,7 @@
  */
 size_t binary_tree_depth(const binary_tree_t *tree)
 {
-	unsigned int depth  = 0;
+	size_t depth  = 0;
 
 	if (tree && tree->parent)
 	{
